requirements.txt
Flask-JWT
Flask-JWT-Extended
Flask-Restful
Flask-SQLAlchemy
Flask-marshmallow
marshmallow-SQLAlchemy
flask-cors
pyodbc